Client/requirements.txt
Pillow==9.2.0
Pillow-PIL==0.1.dev0
pynput==1.7.6
six==1.16.0
